+++ full/src/main/java/org/apache/metamodel/DataContextFactory.java
@@ -31,7 +31,8 @@ import org.apache.metamodel.cassandra.CassandraDataContext;
 import org.apache.metamodel.couchdb.CouchDbDataContext;
 import org.apache.metamodel.csv.CsvConfiguration;
 import org.apache.metamodel.csv.CsvDataContext;
-import org.apache.metamodel.elasticsearch.ElasticSearchDataContext;
+import org.apache.metamodel.elasticsearch.nativeclient.ElasticSearchDataContext;
+import org.apache.metamodel.elasticsearch.rest.ElasticSearchRestDataContext;
 import org.apache.metamodel.excel.ExcelConfiguration;
 import org.apache.metamodel.excel.ExcelDataContext;
 import org.apache.metamodel.fixedwidth.FixedWidthConfiguration;
@@ -56,6 +57,8 @@ import com.mongodb.DB;
 import com.mongodb.MongoClient;
 import com.mongodb.MongoCredential;
 import com.mongodb.ServerAddress;
+
+import io.searchbox.client.JestClient;
 
 /**
  * A factory for DataContext objects. This class substantially easens the task
@@ -147,9 +150,6 @@ public class DataContextFactory {
 
     /**
      * Creates a DataContext based on a JSON file
-     * 
-     * @param file
-     * @return
      */
     public static DataContext createJsonDataContext(File file) {
         return new JsonDataContext(file);
@@ -649,6 +649,18 @@ public class DataContextFactory {
             return new CouchDbDataContext(httpClientBuilder);
         }
         return new CouchDbDataContext(httpClientBuilder, tableDefs);
+    }
+
+    /**
+     * Creates a new JSON-based ElasticSearch datacontext.
+     * @param client
+     *       The Jest client
+     * @param indexName
+     *       The ElasticSearch index name
+     * @return a DataContext object that matches the request
+     */
+    public static UpdateableDataContext createElasticSearchDataContext(JestClient client, String indexName) {
+        return new ElasticSearchRestDataContext(client, indexName);
     }
 
     /**