+++ CHANGES.md
@@ -1,3 +1,10 @@
+### Apache MetaModel (work-in-progress)
+
+ * [METAMODEL-212] - New module for ElasticSearch via REST client.
+ * [METAMODEL-207] - Ensured the serializability of the SingleLineCsvRow class.
+ * [METAMODEL-211] - Fixed a bug related to lookup by primary key (_id) on MongoDB.
+ * [METAMODEL-15] - Query parser support for table names with space. Delimitters can be double quote or square brackets. 
+
 ### Apache MetaModel 4.4.1
 
  * [METAMODEL-198] - Fixed support for JDBC TIMESTAMP precision to match the underlying database's precision.